complete rewrite

+++ src/RealFileSystem.php
@@ -1,31 +1,42 @@
 <?php
 
-namespace ddruganov\TransactionFs;
+namespace ddruganov\TransactionalFileSystem;
 
-use ddruganov\TransactionFs\helpers\PathHelper;
+use ddruganov\TransactionalFileSystem\common\FileSystemInterface;
+use ddruganov\TransactionalFileSystem\common\FileSystemUnitStatus;
+use ddruganov\TransactionalFileSystem\exceptions\FileNotFoundException;
+use ddruganov\TransactionalFileSystem\exceptions\FolderNotFoundException;
+use ddruganov\TransactionalFileSystem\helpers\PathHelper;
+use RecursiveDirectoryIterator;
+use RecursiveIteratorIterator;
 
 final class RealFileSystem implements FileSystemInterface
 {
     # File
 
-    public function fileExists(string $path): bool
+    public function getFileStatus(string $path): FileSystemUnitStatus
     {
-        return file_exists($this->preparePath($path));
+        if (file_exists($this->preparePath($path))) {
+            return FileSystemUnitStatus::EXISTS;
+        }
+
+        return FileSystemUnitStatus::NOT_FOUND;
     }
 
     public function readFile(string $path): ?string
     {
         $path = $this->preparePath($path);
-        return file_exists($path)
-            ? file_get_contents($path)
-            : null;
+        if (!file_exists($path)) {
+            throw new FileNotFoundException($path);
+        }
+        return file_get_contents($path);
     }
 
-    public function writeFile(string $path, string $content, bool $append = false): bool
+    public function writeFile(string $path, string $content = '', bool $append = false): bool
     {
         $path = $this->preparePath($path);
-        if (!file_exists($path)) {
-            $this->createFolder(PathHelper::getFolder($path));
+        if (!file_exists(PathHelper::getFolder($path))) {
+            return false;
         }
         return file_put_contents($path, $content, ($append ? FILE_APPEND : 0)) !== false;
     }
@@ -38,29 +49,47 @@ final class RealFileSystem implements FileSystemInterface
 
     # Folder
 
-    public function folderExists(string $path): bool
+    public function getFolderStatus(string $path): FileSystemUnitStatus
     {
-        return file_exists($this->preparePath($path));
+        if (file_exists($this->preparePath($path))) {
+            return FileSystemUnitStatus::EXISTS;
+        }
+
+        return FileSystemUnitStatus::NOT_FOUND;
     }
 
     public function readFolder(string $path): ?array
     {
         $path = $this->preparePath($path);
-        return file_exists($path)
-            ? array_values(array_filter(scandir($path), fn (string $file) => !in_array($file, ['.', '..'])))
-            : null;
+        if (!file_exists($path)) {
+            throw new FolderNotFoundException($path);
+        }
+
+        $contents = scandir($path);
+        $contents = array_filter($contents, fn (string $file) => !in_array($file, ['.', '..']));
+        return array_values($contents);
     }
 
     public function createFolder(string $path): bool
     {
-        $path = $this->preparePath($path);
-        return file_exists($path) || mkdir($path, recursive: true);
+        return mkdir($this->preparePath($path), recursive: true);
     }
 
     public function deleteFolder(string $path): bool
     {
         $path = $this->preparePath($path);
-        return file_exists($path) && unlink($path);
+        $files = new RecursiveIteratorIterator(
+            new RecursiveDirectoryIterator($path, RecursiveDirectoryIterator::SKIP_DOTS),
+            RecursiveIteratorIterator::CHILD_FIRST
+        );
+
+        foreach ($files as $fileinfo) {
+            $fileinfo->isDir()
+                ? rmdir($fileinfo->getRealPath())
+                : unlink($fileinfo->getRealPath());
+        }
+
+        return rmdir($path);
     }
 
     private function preparePath(string $path): string
